requirements.txt
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
python-dotenv
pip~=24.0
